[SDK #68] * Move content functionality to endpoints module
