+++ src/griddy/nfl/endpoints/pro/games.py
@@ -6,7 +6,6 @@ from griddy.nfl.endpoints.pro import ProSDK
 from griddy.nfl.endpoints.pro.mixins import GameContentMixin, GameScheduleMixin
 from griddy.nfl.types import UNSET, OptionalNullable
 from griddy.nfl.utils import get_security_from_env
-from griddy.nfl.utils.unmarshal_json_response import unmarshal_json_response
 
 
 class Games(ProSDK, GameScheduleMixin, GameContentMixin):
